refactor(utils): utilsモジュールを機能別ヘルパーに分割 (#480)
- HTTP共通ヘルパーを utils/http_helpers.py に抽出（設定解決、エラーマッピング、ログ）
- 非同期基底クライアントを utils/jsonplaceholder_base_async.py に作成
- 同期基底クライアントを utils/jsonplaceholder_base_sync.py に作成
- レスポンス解析ユーティリティを utils/response_parsing.py に追加
- リトライロジック（指数バックオフ+jitter）を utils/retry.py に追加
- api_client.py と github_client.py を新モジュール使用に更新
- PR#347 の DRY 原則（Q-8: インライン重複防止）に対応

refactor(utils): Split utils into focused modules (PR#461)
