+++ utils/api_client.py
@@ -1,29 +1,46 @@
 """同期・非同期HTTPAPIクライアント"""
 
 import asyncio
-import json
-import random
-import sys
-import time
-from types import TracebackType
-from typing import Any, Final, Self, TypedDict
-
-import httpx
-from pydantic import BaseModel, TypeAdapter, ValidationError
-from structlog.typing import FilteringBoundLogger
+from typing import Any, Final, TypedDict
 
 from config.settings import settings
 from models.responses import Album, Comment, Photo, Post, Todo, User
 from utils.exceptions import (
     ASYNC_FATAL_EXCEPTIONS,
-    APIClientError,
-    APIConnectionError,
-    APIHTTPError,
-    APIJSONDecodeError,
-    APIRetryError,
-    APITimeoutError,
 )
-from utils.logger import get_logger
+from utils.exceptions import (
+    APIClientError as APIClientError,
+)
+from utils.exceptions import (
+    APIConnectionError as APIConnectionError,
+)
+from utils.exceptions import (
+    APIHTTPError as APIHTTPError,
+)
+from utils.exceptions import (
+    APIJSONDecodeError as APIJSONDecodeError,
+)
+from utils.exceptions import (
+    APIRetryError as APIRetryError,
+)
+from utils.exceptions import (
+    APITimeoutError as APITimeoutError,
+)
+from utils.http_helpers import (
+    classify_error,
+    log_error_with_stderr_fallback,
+    map_request_error,
+    resolve_client_config,
+    validate_optional_int,
+)
+from utils.jsonplaceholder_base_async import AsyncAPIClient
+from utils.jsonplaceholder_base_sync import SyncAPIClient
+from utils.response_parsing import (
+    parse_response_model,
+    parse_response_model_list,
+    safe_parse_json,
+)
+from utils.retry import exponential_backoff_with_jitter as exponential_backoff_with_jitter
 
 SYNC_FATAL_EXCEPTIONS: tuple[type[BaseException], ...] = (
     KeyboardInterrupt,
@@ -32,572 +49,15 @@ SYNC_FATAL_EXCEPTIONS: tuple[type[BaseException], ...] = (
     RecursionError,
 )
 
-
-def exponential_backoff_with_jitter(
-    attempt: int,
-    base_delay: float = 1.0,
-    max_delay: float = 60.0,
-    jitter_percent: float = 0.3,
-) -> float:
-    """指数バックオフ + 30%ジッター計算
-
-    Args:
-        attempt: 現在のリトライ回数（0始まり）
-        base_delay: 基本遅延時間（秒）
-        max_delay: 最大遅延時間（秒）
-        jitter_percent: ジッター率（デフォルト30%）
-
-    Returns:
-        計算された遅延時間（秒、最小0.1秒）
-
-    """
-    # 指数バックオフ計算
-    delay = min(base_delay * (2**attempt), max_delay)
-    # ±30%のジッター追加（リトライ間隔用、暗号用途ではない）
-    jitter = delay * jitter_percent
-    delay = delay + random.uniform(-jitter, jitter)  # noqa: S311
-    # 最小値保証
-    return max(0.1, delay)
-
-
-def _validate_optional_int(value: int | None, name: str, min_value: int) -> None:
-    """オプショナルな整数パラメータの最小値バリデーション。
-
-    Args:
-        value: 検証対象の値（Noneの場合はスキップ）
-        name: パラメータ名（エラーメッセージ用）
-        min_value: 最小許容値（含む）
-
-    Raises:
-        ValueError: valueがmin_valueより小さい場合
-    """
-    if value is not None and value < min_value:
-        raise ValueError(f"{name} must be >= {min_value}")
-
-
-def _safe_parse_json(response: httpx.Response) -> Any:
-    """レスポンスJSONを安全にパース
-
-    Args:
-        response: HTTPレスポンスオブジェクト
-
-    Returns:
-        パースされたJSONデータ
-
-    Raises:
-        APIJSONDecodeError: JSONパース失敗時
-
-    Notes:
-        ``json.JSONDecodeError`` の ``str(e)`` にはパース位置情報（行番号・
-        文字位置）が含まれるが、httpx例外と異なりホスト名・プロキシ設定等の
-        機密情報は含まれない。デバッグに有用な診断情報を保持するため
-        ``str(e)`` をそのまま使用する。
-        また、``APIJSONDecodeError.response`` に元の ``httpx.Response``
-        オブジェクトを保持するが、レスポンスボディには機密データが含まれる
-        可能性があるためログには出力しない。デバッグ時は呼び出し元で
-        ``e.response`` を通じてアクセス可能。
-
-    """
-    try:
-        return response.json()
-    except json.JSONDecodeError as e:
-        raise APIJSONDecodeError(
-            f"Failed to parse JSON response: {e}",
-            response=response,
-        ) from e
-
-
-def _format_validation_error(e: ValidationError) -> str:
-    details = "; ".join(
-        f"{'.'.join(map(str, err.get('loc', ()))) or '<root>'}: "
-        f"{err.get('msg', 'validation error')} ({err.get('type', 'unknown')})"
-        for err in e.errors(include_input=False)[:3]
-    )
-    more = e.error_count() - 3
-    suffix = f"; ... +{more} more" if more > 0 else ""
-    return f"{e.error_count()} validation error(s): {details}{suffix}"
-
-
-def _parse_response_model[ResponseModelT: BaseModel](
-    response: httpx.Response, model_type: type[ResponseModelT]
-) -> ResponseModelT:
-    """レスポンスJSONをPydanticモデルへ検証して変換する。"""
-    data = _safe_parse_json(response)
-    if not isinstance(data, dict):
-        raise APIJSONDecodeError(
-            f"Expected object JSON for {model_type.__name__}, got {type(data).__name__}",
-            response=response,
-        )
-    try:
-        return model_type.model_validate(data)
-    except ValidationError as e:
-        raise APIJSONDecodeError(
-            f"Invalid {model_type.__name__} response schema: {_format_validation_error(e)}",
-            response=response,
-        ) from e
-
-
-def _parse_response_model_list[ResponseModelT: BaseModel](
-    response: httpx.Response, model_type: type[ResponseModelT]
-) -> list[ResponseModelT]:
-    """レスポンスJSON配列をPydanticモデル配列へ検証して変換する。"""
-    data = _safe_parse_json(response)
-    if not isinstance(data, list):
-        raise APIJSONDecodeError(
-            f"Expected array JSON for {model_type.__name__}, got {type(data).__name__}",
-            response=response,
-        )
-    try:
-        # TypeAdapter(list[model]) を使うと ValidationError の loc に
-        # 失敗要素の index が自動付与される（例: loc=("0", "user_id")）。
-        # _format_validation_error が loc を "." 結合するため "0.user_id: ..."
-        # のように、配列内のどの要素が失敗したか診断可能になる。
-        # NOTE: model_type は実行時には具象クラスだが、mypy は変数を型添字
-        #   list[...] に使えない（valid-type）。実行時の正しさはテストで担保済みのため
-        #   この行に限り type: ignore を付与する（Pydantic + mypy の既知の制約）。
-        return TypeAdapter(list[model_type]).validate_python(data)  # type: ignore[valid-type]
-    except ValidationError as e:
-        raise APIJSONDecodeError(
-            f"Invalid {model_type.__name__} response schema: {_format_validation_error(e)}",
-            response=response,
-        ) from e
-
-
-def _map_request_error(e: httpx.RequestError | httpx.InvalidURL) -> APIClientError:
-    """httpxネットワーク例外をカスタム例外にマッピング
-
-    Args:
-        e: httpx.RequestError または httpx.InvalidURL（またはそのサブクラス）
-
-    Returns:
-        APIClientErrorサブクラス（リトライ可能エラーの場合のみ。
-        非リトライ時（TooManyRedirects / InvalidURL）は
-        APIClientError 基底クラスを raise するため返らない）。
-
-    Raises:
-        APIClientError: 非リトライ可能エラー（TooManyRedirects, InvalidURL）
-
-    Note:
-        httpx例外の扱い:
-        - RequestError サブクラス（リトライ可能）:
-          TimeoutException (ConnectTimeout, ReadTimeout),
-          NetworkError (ConnectError, ReadError, WriteError 等のサブクラスを含む)
-          ※ ConnectError は NetworkError のサブクラスだが個別分岐で処理
-        - RequestError サブクラス（非リトライ）:
-          TooManyRedirects → 即座にraise
-        - 独立例外（RequestError のサブクラスではない、非リトライ）:
-          InvalidURL → 即座にraise
-        生成される例外メッセージは固定プレフィックスと ``type(e).__name__``
-        （例外クラス名）のみで構成され、``str(e)`` は含めない。
-        非リトライエラーは ``raise ... from e`` により即座にスローされる。
-        リトライ可能エラーは ``__cause__ = e`` 手動設定後に返され、
-        呼び出し元で ``raise`` された際にトレースバックで確認できる。
-        なお、``__cause__`` に保持される httpx 例外の文字列には
-        ホスト名・プロキシ設定等の機密情報が含まれることがあり、
-        ``traceback.print_exception(chain=True)``（デフォルト）では
-        この ``__cause__`` チェーンが展開されるため、表示用途では
-        ``chain=False`` を指定して機密漏洩を防ぐこと。
-        ``main()`` で受け取る ``e`` は ``APIClientError`` なので、
-        ``chain=False`` が抑止するのは ``__cause__`` 側の httpx 例外チェーンであり、
-        ``APIClientError`` 本体のスタックトレースは引き続き表示される。
-        参照: ``main()`` の ``traceback.print_exception(e, chain=False)`` 実装。
-
-    """
-    # Non-retryable errors - raise immediately (no point in retrying)
-    if isinstance(e, httpx.TooManyRedirects | httpx.InvalidURL):
-        raise APIClientError(f"Non-retryable request error: {type(e).__name__}") from e
-
-    # Retryable errors: returnするため `raise ... from e` は使えず __cause__ を手動設定する。
-    # PEP 3134: exc.__cause__ = e を設定すると __suppress_context__ が自動で True になり、
-    # 呼び出し元が raise した際に `raise exc from e` と同じ例外チェーン表示になる。
-    if isinstance(e, httpx.TimeoutException):
-        timeout_exc = APITimeoutError(f"Request timeout: {type(e).__name__}")
-        timeout_exc.__cause__ = e
-        return timeout_exc
-    if isinstance(e, httpx.ConnectError):
-        connect_exc = APIConnectionError(f"Connection failed: {type(e).__name__}")
-        connect_exc.__cause__ = e
-        return connect_exc
-    # NetworkError, etc. - retryable network issues
-    network_exc = APIConnectionError(f"Network error: {type(e).__name__}")
-    network_exc.__cause__ = e
-    return network_exc
-
-
-def _resolve_client_config(
-    base_url: str | None,
-    timeout: float | None,
-    retry_count: int | None,
-    retry_delay: float | None,
-    headers: dict[str, str] | None,
-) -> tuple[str, float, int, float, dict[str, str]]:
-    """Sync/Async共通の設定解決ロジック。
-
-    引数またはsettingsから設定値を解決し、バリデーションを実行する。
-    HTTPクライアント初期化・ロガー初期化は呼び出し元の責務。
-
-    Args:
-        base_url: APIのベースURL（Noneの場合settings.api.base_urlを使用）
-        timeout: タイムアウト秒数（Noneの場合settings.api.timeoutを使用）
-        retry_count: リトライ回数（Noneの場合settings.api.retry_countを使用）
-        retry_delay: リトライ間隔秒数（Noneの場合settings.api.retry_delayを使用）
-        headers: 追加ヘッダー（デフォルトヘッダーにマージ）
-
-    Returns:
-        (base_url, timeout, retry_count, retry_delay, default_headers) のタプル
-
-    Raises:
-        ValueError: base_urlが空文字列またはホワイトスペース
-            （str.strip() で除去される文字）のみの文字列の場合
-
-    """
-    base_url = base_url if base_url is not None else settings.api.base_url
-    if not base_url.strip():
-        raise ValueError("base_url が空です。引数または API__BASE_URL 環境変数を確認してください。")
-    timeout = timeout if timeout is not None else settings.api.timeout
-    retry_count = retry_count if retry_count is not None else settings.api.retry_count
-    retry_delay = retry_delay if retry_delay is not None else settings.api.retry_delay
-
-    default_headers = {
-        "User-Agent": settings.api.user_agent,
-        "Accept": "application/json",
-        "Content-Type": "application/json",
-    }
-    # `if headers:` ではなく `is not None` を使用: 空辞書({})を渡した場合も
-    # update()を実行する（no-opだが、Noneと空辞書の意味論を明確に区別するため）
-    if headers is not None:
-        default_headers.update(headers)
-
-    return (
-        base_url,
-        timeout,
-        retry_count,
-        retry_delay,
-        default_headers,
-    )
-
-
-def _classify_error(
-    e: httpx.RequestError | httpx.InvalidURL,
-    logger: FilteringBoundLogger,
-    *,
-    is_async: bool,
-    method: str,
-    endpoint: str,
-) -> APIClientError:
-    """Sync/Async共通のネットワークエラー分類・ログ出力。
-
-    エラー種別に応じてERROR/WARNINGログを出力し、_map_request_error()を呼び出す。
-    非リトライエラー(TooManyRedirects/InvalidURL)は_map_request_error()内で即座にraiseされる。
-
-    Args:
-        e: httpxのリクエストエラーまたはInvalidURL
-        logger: structlogロガーインスタンス
-        is_async: 非同期クライアントからの呼び出しかどうか
-        method: HTTPメソッド名
-        endpoint: APIエンドポイント
-
-    Returns:
-        APIClientErrorサブクラス（リトライ可能エラーの場合）。
-        TooManyRedirects / InvalidURL の場合は _map_request_error() 内で
-        raise されるため、呼び出し元には値が返らない。
-
-    Raises:
-        APIClientError: TooManyRedirects または InvalidURL の場合
-            （logger.error でログ出力後、_map_request_error() を経由して raise される）。
-            注: サブクラスではなく APIClientError 基底クラスが raise される。
-            リトライ可能エラーは logger.warning でログ出力し、raise されない。
-
-    Notes:
-        本関数のログ出力（``request_error_non_retryable`` / ``request_error`` イベント）では
-        ``error`` フィールドを省略している。httpx 例外の文字列には
-        ホスト名、プロキシ設定等の機密情報が含まれるため、``error_type``
-        （例外クラス名）のみ記録してエラー分類に必須情報を確保する。
-        非リトライエラー（``request_error_non_retryable``）は ``logger.error``、
-        リトライ可能エラー（``request_error``）は ``logger.warning`` でログ出力される。
-        例外の生成・チェーン設定の詳細は ``_map_request_error()`` 参照。
-
-    """
-    if isinstance(e, httpx.TooManyRedirects | httpx.InvalidURL):
-        logger.error(
-            "request_error_non_retryable",
-            is_async=is_async,
-            method=method,
-            endpoint=endpoint,
-            error_type=type(e).__name__,
-        )
-    else:
-        logger.warning(
-            "request_error",
-            is_async=is_async,
-            method=method,
-            endpoint=endpoint,
-            error_type=type(e).__name__,
-        )
-    return _map_request_error(e)
-
-
-# =============================================================================
-# 基本HTTPクライアント
-# =============================================================================
-
-
-class SyncAPIClient:
-    """基本的な同期HTTPクライアント"""
-
-    def __init__(
-        self,
-        base_url: str | None = None,
-        timeout: float | None = None,
-        retry_count: int | None = None,
-        retry_delay: float | None = None,
-        headers: dict[str, str] | None = None,
-    ):
-        """Args:
-        base_url: APIのベースURL（設定から自動取得可能）
-        timeout: リクエストタイムアウト（秒）
-        retry_count: リトライ回数
-        retry_delay: リトライ間隔（秒）
-        headers: 追加HTTPヘッダー
-
-        Raises:
-            ValueError: base_urlが空文字列またはホワイトスペース
-                （スペース・タブ・改行等）のみの文字列の場合
-
-        """
-        # 設定解決・バリデーション（Sync/Async共通ロジック）
-        # NOTE: retry_count=0, retry_delay=0.0, timeout=0.0 は有効な設定値のため is not None で判定
-        # timeout=0.0: 即座にタイムアウト（無効化は timeout=None）
-        (
-            self.base_url,
-            self.timeout,
-            self.retry_count,
-            self.retry_delay,
-            self.default_headers,
-        ) = _resolve_client_config(base_url, timeout, retry_count, retry_delay, headers)
-
-        # ロガーの初期化（structlog統合）
-        self.logger = get_logger(__name__)
-
-        # HTTPクライアントの初期化
-        # close() 後に None を代入するため Optional 宣言
-        # （AsyncAPIClient._client と対称, PR#347 CQ-6）。
-        self._client: httpx.Client | None = httpx.Client(
-            base_url=self.base_url,
-            timeout=self.timeout,
-            headers=self.default_headers,
-            limits=httpx.Limits(max_connections=settings.api.max_connections),
-        )
-
-        self.logger.info("api_client_initialized", base_url=self.base_url)
-
-    def __enter__(self) -> Self:
-        """コンテキストマネージャーのエントリー"""
-        return self
-
-    def __exit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc_val: BaseException | None,
-        exc_tb: TracebackType | None,
-    ) -> None:
-        """コンテキストマネージャーの終了処理"""
-        self.close()
-
-    def close(self) -> None:
-        """クライアントのクローズ
-
-        ``AsyncAPIClient`` / ``AsyncGitHubClient.__aexit__`` と同様に、close 後は
-        ``self._client = None`` を設定してダブルクローズを防止する（PR#347 CQ-6）。
-        truthy チェックではなく ``is not None`` で他箇所の規約と統一する。
-
-        ``self._client.close()`` が例外（``httpx.CloseError`` / ``OSError`` 等）を投げても、
-        ``finally`` 節で ``self._client = None`` を必ず設定する。これにより ``_request`` 冒頭の
-        use-after-close ガード（``_client is None`` 判定）の前提が保たれ、close 失敗後に壊れた
-        クライアントへリクエストが発行される状態不整合を防ぐ。``AsyncAPIClient._close_async_client``
-        が全 except 経路で ``_client=None`` を設定するのと対称（PR#347 review）。例外は従来通り
-        呼び出し元へ伝播させる（``finally`` は抑制しない）。``api_client_closed`` info ログは
-        close 成功時のみ出力する（``finally`` の外に置くため、例外時はスキップされる）。
-        """
-        if self._client is not None:
-            try:
-                self._client.close()
-            finally:
-                self._client = None
-            self.logger.info("api_client_closed")
-
-    def _make_request_with_retry(self, method: str, endpoint: str, **kwargs: Any) -> httpx.Response:
-        """リトライ機能付きHTTPリクエスト実行
-
-        Args:
-            method: HTTPメソッド
-            endpoint: APIエンドポイント
-            **kwargs: httpxに渡す追加パラメータ
-
-        Returns:
-            httpx.Response: APIレスポンス
-
-        Raises:
-            APIConnectionError: 接続エラー
-            APITimeoutError: タイムアウトエラー
-            APIHTTPError: HTTPステータスエラー
-            APIRetryError: リトライ上限エラー
-            APIClientError: 非リトライエラー（TooManyRedirects / InvalidURL）
-
-        Note:
-            TooManyRedirects/InvalidURL は _map_request_error() 内で即 raise されるため、
-            APIRetryError ではなく APIClientError として呼び出し元に届く。
-            呼び出し元は APIClientError で捕捉すること。
-
-        """
-        # close 後の use-after-close を明示エラー化（AsyncAPIClient._request と同一パターン）。
-        # 型注釈 _client: httpx.Client | None に対する None 絞り込みも兼ねる（PR#347 CQ-6）。
-        if self._client is None:
-            raise RuntimeError("Client not initialized or already closed.")
-
-        last_exception: APIClientError | None = None
-
-        for attempt in range(self.retry_count + 1):
-            # HTTPリクエスト実行（ネットワーク層）
-            try:
-                # structlogでログ出力（DRY原則: 重複ログ削除）
-                if attempt > 0:
-                    self.logger.warning(
-                        "request_retry",
-                        attempt=attempt + 1,
-                        max_attempts=self.retry_count + 1,
-                        method=method,
-                        endpoint=endpoint,
-                    )
-                else:
-                    self.logger.debug("request_start", method=method, endpoint=endpoint)
-
-                # HTTPリクエスト実行
-                response = self._client.request(method, endpoint, **kwargs)
-            except (httpx.RequestError, httpx.InvalidURL) as e:
-                # 全ネットワーク層エラーをキャッチ（TimeoutException, ConnectError, etc.）
-                # TooManyRedirects/InvalidURL は _classify_error → _map_request_error 内で即 raise
-                last_exception = _classify_error(
-                    e,
-                    self.logger,
-                    is_async=False,
-                    method=method,
-                    endpoint=endpoint,
-                )
-            else:
-                # ネットワーク成功時のみHTTPステータス処理
-                try:
-                    response.raise_for_status()
-                    self.logger.debug(
-                        "request_success",
-                        method=method,
-                        endpoint=endpoint,
-                        status_code=response.status_code,
-                    )
-                    return response
-                except httpx.HTTPStatusError as e:
-                    # 4xxエラーはリトライしない（クライアントエラー）
-                    if e.response.is_client_error:
-                        self.logger.error(
-                            "client_error",
-                            status_code=e.response.status_code,
-                            method=method,
-                            endpoint=endpoint,
-                        )
-                        raise APIHTTPError(
-                            f"HTTP {e.response.status_code} Client Error",
-                            e.response.status_code,
-                            e.response,
-                        ) from e
-
-                    # 5xxエラーはリトライ対象
-                    self.logger.warning(
-                        "server_error",
-                        status_code=e.response.status_code,
-                        method=method,
-                        endpoint=endpoint,
-                    )
-                    last_exception = APIHTTPError(
-                        f"HTTP {e.response.status_code} Server Error",
-                        e.response.status_code,
-                        e.response,
-                    )
-
-            # 最後の試行でなければ指数バックオフ + 30%ジッターで待機
-            if attempt < self.retry_count:
-                delay = exponential_backoff_with_jitter(
-                    attempt=attempt,
-                    base_delay=self.retry_delay,
-                    jitter_percent=0.3,
-                )
-                self.logger.debug(
-                    "retry_backoff",
-                    delay_seconds=round(delay, 2),
-                    attempt=attempt + 1,
-                    strategy="exponential_backoff_with_jitter",
-                )
-                time.sleep(delay)
-
-        # すべてのリトライが失敗
-        self.logger.error("all_retries_failed", method=method, endpoint=endpoint)
-        raise APIRetryError(
-            f"Request failed after {self.retry_count + 1} attempts",
-        ) from last_exception
-
-    def get(
-        self,
-        endpoint: str,
-        params: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """GETリクエスト実行"""
-        return self._make_request_with_retry("GET", endpoint, params=params, headers=headers)
-
-    def post(
-        self,
-        endpoint: str,
-        json: dict[str, Any] | None = None,
-        data: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """POSTリクエスト実行"""
-        return self._make_request_with_retry(
-            "POST",
-            endpoint,
-            json=json,
-            data=data,
-            headers=headers,
-        )
-
-    def put(
-        self,
-        endpoint: str,
-        json: dict[str, Any] | None = None,
-        data: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """PUTリクエスト実行"""
-        return self._make_request_with_retry("PUT", endpoint, json=json, data=data, headers=headers)
-
-    def delete(self, endpoint: str, headers: dict[str, str] | None = None) -> httpx.Response:
-        """DELETEリクエスト実行"""
-        return self._make_request_with_retry("DELETE", endpoint, headers=headers)
-
-    def patch(
-        self,
-        endpoint: str,
-        json: dict[str, Any] | None = None,
-        data: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """PATCHリクエスト実行"""
-        return self._make_request_with_retry(
-            "PATCH",
-            endpoint,
-            json=json,
-            data=data,
-            headers=headers,
-        )
+# Backward-compatible aliases retained until W3 import cutover.
+_validate_optional_int = validate_optional_int
+_safe_parse_json = safe_parse_json
+_parse_response_model = parse_response_model
+_parse_response_model_list = parse_response_model_list
+_map_request_error = map_request_error
+_resolve_client_config = resolve_client_config
+_classify_error = classify_error
+_log_error_with_stderr_fallback = log_error_with_stderr_fallback
 
 
 # =============================================================================
@@ -802,434 +262,9 @@ class SyncJSONPlaceholderClient(SyncAPIClient):
             return False
 
 
-def _log_error_with_stderr_fallback(
-    logger: FilteringBoundLogger,
-    source: str,
-    context: str,
-    exc: BaseException,
-    event: str,
-    **fields: Any,
-) -> None:
-    """logger.error 記録 + 失敗時 stderr フォールバック (PR#347 B-3 / Q-8 DRY)。
-
-    ``api_client`` / ``github_client`` の close・cache 失敗ログで共通する
-    「``logger.error`` → 失敗時 ``stderr``」パターンをモジュールレベルに集約する。
-    ``AsyncGitHubClient`` は ``AsyncAPIClient`` を継承しないため、メソッドではなく
-    モジュール関数として共有する (PR#347 Q-8: インライン重複による修正漏れを防ぐ)。
-
-    ロガー自体が致命例外 (``MemoryError`` / ``RecursionError``) を投げた場合は
-    fail-fast で再 raise し、それ以外のロガー例外は握りつぶした ``exc`` の型名を
-    ``stderr`` へ再露出させて監視可能性を保つ。
-
-    Args:
-        logger: structlog ロガー。
-        source: stderr メッセージのソース識別子 (例 ``"api_client"``)。
-        context: 失敗箇所の短い識別子 (例 ``"aclose"`` / ``"etag_cache"``)。
-        exc: 既に握りつぶされている元例外 (型名のみ stderr 出力, PII 非含)。
-        event: ``logger.error`` へ渡すイベント名。
-        **fields: ``logger.error`` へ渡す構造化フィールド。
-
-    Raises:
-        MemoryError: ``logger`` が ``MemoryError`` を投げた場合に再 raise（fail-fast）。
-        RecursionError: ``logger`` が ``RecursionError`` を投げた場合に再 raise（fail-fast）。
-    """
-    try:
-        logger.error(event, **fields)
-    except (MemoryError, RecursionError):  # fmt: skip
-        # 致命例外は握りつぶさず再raise（fail-fast）。両者は Exception 派生のため、
-        # 下の except Exception より先に明示的に先取りする
-        # （sentry_init._safe_log_warning / _close_async_client と同一方針）。
-        raise
-    except Exception:  # noqa: BLE001
-        # ロガー例外が握りつぶした exc を再露出させない保険。
-        try:
-            print(
-                f"[{source}] {context} logger failed: {type(exc).__name__}",
-                file=sys.stderr,
-                flush=True,
-            )
-        except Exception:  # noqa: BLE001, S110
-            pass
-
-
 # =============================================================================
-# 非同期APIクライアント
+# JSONPlaceholder Async APIクライアント
 # =============================================================================
-
-
-class AsyncAPIClient:
-    """非同期HTTPクライアント"""
-
-    _client: httpx.AsyncClient | None  # aclose() 後に None を代入するため明示宣言
-
-    def __init__(
-        self,
-        base_url: str | None = None,
-        timeout: float | None = None,
-        retry_count: int | None = None,
-        retry_delay: float | None = None,
-        headers: dict[str, str] | None = None,
-    ):
-        """Args:
-        base_url: APIのベースURL（設定から自動取得可能）
-        timeout: リクエストタイムアウト（秒）
-        retry_count: リトライ回数
-        retry_delay: リトライ間隔（秒）
-        headers: 追加HTTPヘッダー
-
-        Raises:
-            ValueError: base_urlが空文字列またはホワイトスペース
-                （スペース・タブ・改行等）のみの文字列の場合
-
-        """
-        # 設定解決・バリデーション（Sync/Async共通ロジック）
-        # NOTE: retry_count=0, retry_delay=0.0, timeout=0.0 は有効な設定値のため is not None で判定
-        # timeout=0.0: 即座にタイムアウト（無効化は timeout=None）
-        (
-            self.base_url,
-            self.timeout,
-            self.retry_count,
-            self.retry_delay,
-            self.default_headers,
-        ) = _resolve_client_config(base_url, timeout, retry_count, retry_delay, headers)
-
-        # ロガーの初期化（structlog統合）
-        self.logger = get_logger(__name__)
-
-        # HTTPクライアントの初期化（非同期）
-        self._client = httpx.AsyncClient(
-            base_url=self.base_url,
-            timeout=self.timeout,
-            headers=self.default_headers,
-            limits=httpx.Limits(max_connections=settings.api.max_connections),
-        )
-
-        self.logger.info("async_api_client_initialized", base_url=self.base_url)
-
-    async def __aenter__(self) -> Self:
-        """非同期コンテキストマネージャーのエントリー"""
-        return self
-
-    def _log_aclose_error_with_fallback(
-        self, event: str, close_exc: BaseException, **fields: Any
-    ) -> None:
-        """aclose エラーを記録し、失敗時は stderr へフォールバック (PR#347 B-3 / Q-8)。
-
-        モジュールレベル ``_log_error_with_stderr_fallback`` への薄いラッパー。
-        ``github_client`` とロジック (logger.error → 失敗時 stderr) を共有し、
-        インライン重複による修正漏れを防ぐ (PR#347 Q-8 DRY)。
-        """
-        _log_error_with_stderr_fallback(
-            self.logger, "api_client", "aclose", close_exc, event, **fields
-        )
-
-    async def _close_async_client(
-        self,
-        body_exc_type: type[BaseException] | None,
-        *,
-        suppress_unexpected: bool = False,
-    ) -> None:
-        """``__aexit__`` と ``aclose()`` で共有する close 処理.
-
-        Args:
-            body_exc_type: ``__aexit__`` 経路では context manager の body 内で発生した
-                例外型を渡す (例外無しなら ``None``)。``aclose()`` 直接呼び出し経路では
-                常に ``None`` を渡す。``aclose()`` 失敗時に予期しない close 例外
-                (``Exception`` 派生) を捕捉した際、body 例外の上書き防止のため
-                ``has_body_exception = body_exc_type is not None`` を判定材料として
-                利用する。``None`` の場合のみ実装バグとして bare ``raise`` で再送出する。
-            suppress_unexpected: ``True`` の場合、予期しない close 例外を error ログ
-                のみ記録して握りつぶす (re-raise しない)。``aclose()`` 直接呼び出し経路
-                で ``True`` を渡すことで finally ブロック等での安全な呼び出しを保証する。
-                ``__aexit__`` 経路では ``False`` (デフォルト) のまま ``has_body_exception``
-                ロジックによる従来の re-raise 判定を維持する。
-        """
-        if self._client is not None:
-            try:
-                await self._client.aclose()
-            except (httpx.CloseError, OSError) as close_exc:
-                # 既知のクローズ時例外 — 警告のみ（body 例外を上書きしない）
-                self._client = None
-                self.logger.warning(
-                    "async_api_client_aclose_failed",
-                    error_type=type(close_exc).__name__,
-                    error_module=type(close_exc).__module__,
-                )
-            except (MemoryError, RecursionError):  # fmt: skip
-                # MemoryError/RecursionError も Exception 派生のため、再raise しないと
-                # 下流の except Exception に捕捉されサイレント隠蔽される
-                # （github_client / sentry_init と同一方針）。
-                # 致命的エラーとして必ず再raise（fail-fast）。
-                # ASYNC_FATAL_EXCEPTIONS は流用しない（close 文脈で asyncio.CancelledError
-                # は捕捉対象外＝BaseException 直系で素通りさせるのが正しいため）。
-                raise
-            except Exception as close_exc:  # noqa: BLE001
-                # 予期しない例外（AttributeError, RuntimeError, ValueError, TypeError 等の
-                # 実装バグ可能性）を捕捉。以下は本句より先に処理済み / 境界外:
-                #   - RecursionError / MemoryError: 上の専用句で先取り捕捉し
-                #     即時 re-raise（fail-fast）。
-                #   - KeyboardInterrupt / SystemExit / asyncio.CancelledError は
-                #     BaseException 直系で `except Exception` の境界外。
-                #     ユーザー停止/プロセス終了/cancellation を妨げない。
-                has_body_exception = body_exc_type is not None
-                if suppress_unexpected:
-                    # aclose() 直接呼び出し経路: finally ブロック等での安全な呼び出しを保証するため
-                    # 予期しない例外を握りつぶし、error ログで本番監視対象にする。
-                    # suppress_unexpected=True は has_body_exception より優先して評価する。
-                    # suppress 経路でも状態一貫性のため None セット。
-                    # aclose 失敗後の壊れたクライアント再利用を防止
-                    # （github_client __aexit__ L356 / 成功時 else 節と同一方針）。
-                    self._client = None
-                    self._log_aclose_error_with_fallback(
-                        "async_api_client_aclose_unexpected_error_suppressed",
-                        close_exc,
-                        error_type=type(close_exc).__name__,
-                        error_module=type(close_exc).__module__,
-                        exc_info=True,  # スタックトレースをログに残す
-                    )
-                else:
-                    # __aexit__ 経路（context manager）: 従来の has_body_exception ロジックを維持。
-                    # PR#347 review SF-2: close_exc が body 例外を上書きしないため
-                    # __context__ チェーンは切断される。代わりに body 例外の型名を
-                    # 同一ログイベント内に記録し、close 失敗と body 例外の対応関係を
-                    # 追跡可能にする (PII 非含: __qualname__ はクラス名のみ)。
-                    self._log_aclose_error_with_fallback(
-                        "async_api_client_aclose_unexpected_error",
-                        close_exc,
-                        error_type=type(close_exc).__name__,
-                        error_module=type(close_exc).__module__,
-                        has_body_exception=has_body_exception,
-                        action=(
-                            "suppressed_due_to_body_exception"
-                            if has_body_exception
-                            else "re_raised"
-                        ),
-                        body_exception_type=(
-                            body_exc_type.__qualname__ if body_exc_type is not None else None
-                        ),
-                        exc_info=True,  # スタックトレースをログに残す
-                    )
-                    # body 例外がない場合のみ実装バグとして re-raise。
-                    # body 例外がある場合は本質的原因の上書きを防ぐため raise しない。
-                    # bare ``raise`` で active exception の traceback を完全保持
-                    # （``raise close_exc`` への回帰防止: 余分な frame を追加せず Python idiom）。
-                    if not has_body_exception:
-                        raise
-            else:
-                # aclose() 成功時のみ closed ログを出す。logger.info を try 内に置くと
-                # logger 自体の例外が aclose 失敗として誤検知されるため else 節に分離。
-                # github_client.py / Sync close() と同一パターン (PR#347 Q-1 Codex fix)。
-                self._client = None  # double-close 防止（_client 型は | None 宣言済み）
-                self.logger.info("async_api_client_closed")
-
-    async def __aexit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc_val: BaseException | None,
-        exc_tb: TracebackType | None,
-    ) -> None:
-        """非同期コンテキストマネージャーの終了処理"""
-        await self._close_async_client(exc_type)
-
-    async def aclose(self) -> None:
-        """クライアントのクローズ
-
-        async with パターンと aclose() 単独呼び出しの両経路で
-        ``async_api_client_closed`` ログを出力し、Sync (``SyncAPIClient.close()``)
-        との observability 対称性を保つ (PR#347 review Q-1)。
-
-        Note: 直接呼び出し時 (async context manager 経由でない場合)、予期しない close 例外は
-        error ログ記録の上で抑制される (re-raise しない)。これにより finally ブロックでの
-        安全な呼び出しを保証し、アプリクラッシュリスクを回避する。
-        async with 経由では body 例外保護ロジック (has_body_exception) が適用され、
-        body 例外がない場合のみ close 例外を re-raise する従来の挙動を維持する。
-        """
-        await self._close_async_client(None, suppress_unexpected=True)
-
-    async def _make_request_with_retry(
-        self,
-        method: str,
-        endpoint: str,
-        **kwargs: Any,
-    ) -> httpx.Response:
-        """リトライ機能付き非同期HTTPリクエスト実行
-
-        Args:
-            method: HTTPメソッド
-            endpoint: APIエンドポイント
-            **kwargs: httpxに渡す追加パラメータ
-
-        Returns:
-            httpx.Response: APIレスポンス
-
-        Raises:
-            APIConnectionError: 接続エラー
-            APITimeoutError: タイムアウトエラー
-            APIHTTPError: HTTPステータスエラー
-            APIRetryError: リトライ上限エラー
-            APIClientError: 非リトライエラー（TooManyRedirects / InvalidURL）
-
-        Note:
-            TooManyRedirects/InvalidURL は _map_request_error() 内で即 raise されるため、
-            APIRetryError ではなく APIClientError として呼び出し元に届く。
-            呼び出し元は APIClientError で捕捉すること。
-
-        """
-        # close 後の use-after-close を明示エラー化（github_client.py L878 と同一パターン）。
-        # 型注釈 _client: AsyncClient | None に対する None 絞り込みも兼ねる。
-        if self._client is None:
-            raise RuntimeError("Client not initialized. Use 'async with' context.")
-
-        last_exception: APIClientError | None = None
-
-        for attempt in range(self.retry_count + 1):
-            # 非同期HTTPリクエスト実行（ネットワーク層）
-            try:
-                # structlogでログ出力（DRY原則: 重複ログ削除）
-                if attempt > 0:
-                    self.logger.warning(
-                        "async_request_retry",
-                        attempt=attempt + 1,
-                        max_attempts=self.retry_count + 1,
-                        method=method,
-                        endpoint=endpoint,
-                    )
-                else:
-                    self.logger.debug("async_request_start", method=method, endpoint=endpoint)
-
-                # 非同期HTTPリクエスト実行
-                response = await self._client.request(method, endpoint, **kwargs)
-            except (httpx.RequestError, httpx.InvalidURL) as e:
-                # 全ネットワーク層エラーをキャッチ（TimeoutException, ConnectError, etc.）
-                # TooManyRedirects/InvalidURL は _classify_error → _map_request_error 内で即 raise
-                last_exception = _classify_error(
-                    e,
-                    self.logger,
-                    is_async=True,
-                    method=method,
-                    endpoint=endpoint,
-                )
-            else:
-                # ネットワーク成功時のみHTTPステータス処理
-                try:
-                    response.raise_for_status()
-                    self.logger.debug(
-                        "async_request_success",
-                        method=method,
-                        endpoint=endpoint,
-                        status_code=response.status_code,
-                    )
-                    return response
-                except httpx.HTTPStatusError as e:
-                    # 4xxエラーはリトライしない（クライアントエラー）
-                    if e.response.is_client_error:
-                        self.logger.error(
-                            "client_error",
-                            status_code=e.response.status_code,
-                            method=method,
-                            endpoint=endpoint,
-                        )
-                        raise APIHTTPError(
-                            f"HTTP {e.response.status_code} Client Error",
-                            e.response.status_code,
-                            e.response,
-                        ) from e
-
-                    # 5xxエラーはリトライ対象
-                    self.logger.warning(
-                        "server_error",
-                        status_code=e.response.status_code,
-                        method=method,
-                        endpoint=endpoint,
-                    )
-                    last_exception = APIHTTPError(
-                        f"HTTP {e.response.status_code} Server Error",
-                        e.response.status_code,
-                        e.response,
-                    )
-
-            # 最後の試行でなければ指数バックオフ + 30%ジッターで待機
-            if attempt < self.retry_count:
-                delay = exponential_backoff_with_jitter(
-                    attempt=attempt,
-                    base_delay=self.retry_delay,
-                    jitter_percent=0.3,
-                )
-                self.logger.debug(
-                    "async_retry_backoff",
-                    delay_seconds=round(delay, 2),
-                    attempt=attempt + 1,
-                    strategy="exponential_backoff_with_jitter",
-                )
-                await asyncio.sleep(delay)
-
-        # すべてのリトライが失敗
-        self.logger.error("async_all_retries_failed", method=method, endpoint=endpoint)
-        raise APIRetryError(
-            f"Async request failed after {self.retry_count + 1} attempts",
-        ) from last_exception
-
-    async def get(
-        self,
-        endpoint: str,
-        params: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """非同期GETリクエスト実行"""
-        return await self._make_request_with_retry("GET", endpoint, params=params, headers=headers)
-
-    async def post(
-        self,
-        endpoint: str,
-        json: dict[str, Any] | None = None,
-        data: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """非同期POSTリクエスト実行"""
-        return await self._make_request_with_retry(
-            "POST",
-            endpoint,
-            json=json,
-            data=data,
-            headers=headers,
-        )
-
-    async def put(
-        self,
-        endpoint: str,
-        json: dict[str, Any] | None = None,
-        data: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """非同期PUTリクエスト実行"""
-        return await self._make_request_with_retry(
-            "PUT",
-            endpoint,
-            json=json,
-            data=data,
-            headers=headers,
-        )
-
-    async def delete(self, endpoint: str, headers: dict[str, str] | None = None) -> httpx.Response:
-        """非同期DELETEリクエスト実行"""
-        return await self._make_request_with_retry("DELETE", endpoint, headers=headers)
-
-    async def patch(
-        self,
-        endpoint: str,
-        json: dict[str, Any] | None = None,
-        data: dict[str, Any] | None = None,
-        headers: dict[str, str] | None = None,
-    ) -> httpx.Response:
-        """非同期PATCHリクエスト実行"""
-        return await self._make_request_with_retry(
-            "PATCH",
-            endpoint,
-            json=json,
-            data=data,
-            headers=headers,
-        )
-
 
 # 一括作成の部分失敗ログで記録する詳細の上限件数
 _MAX_LOGGED_FAILURE_DETAILS: Final[int] = 5
